moved snapblue files under the src directory
